Refactor the table widget out into a common file
